Start with `eskip.ParseFilters` and fix failing tests logs
