+++ dataclients/kubernetes/testdata/routegroups/convert/no-catchall-for-failed-route-group.log
@@ -1,2 +1,1 @@
-kind=RouteGroup name=myapp
-syntax error
+\[routegroup\] parse failed after token foo, last route id: , position 8: syntax error
